출석 체크 및 수강 신청 수정본
