+++ lib/main.dart
@@ -15,7 +15,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: Attendance(),
+      home: classList(),
     );
   }
 }
