+++ lib/src/page/class_list.dart
@@ -2,8 +2,8 @@ import 'package:checkcheck_project/src/page/registration_page.dart';
 import 'package:checkcheck_project/src/widget/custom_button.dart';
 import 'package:checkcheck_project/src/widget/custom_elevated_botton.dart';
 import 'package:checkcheck_project/src/widget/custom_text_field.dart';
+import 'package:checkcheck_project/src/widget/tip.dart';
 import 'package:flutter/material.dart';
-import 'package:checkcheck_project/src/widget/custom_scrollbar.dart';
 
 class classList extends StatelessWidget {
 
@@ -14,38 +14,39 @@ class classList extends StatelessWidget {
       backgroundColor: Colors.black,
       body: Padding(
         padding: const EdgeInsets.all(8.0),
-        child: CustomScrollbar(
-          // CustomScrollbar 추가
-          child: SingleChildScrollView(
-            // 스크롤 가능한 영역 추가
-            child: Column(
-              children: [
-                const Padding(
-                  padding: EdgeInsets.symmetric(horizontal:  8.0),
-                  child: CustomTextField(
-                    hintText: '검색하기',
-                    suffixIcon: Icon(Icons.search),
+          child: Scrollbar(
+            thumbVisibility: true,
+            child: SingleChildScrollView(
+              child: Column(
+                children: [
+                  const Padding(
+                    padding: EdgeInsets.symmetric(horizontal:  8.0),
+                    child: CustomTextField(
+                      hintText: '검색하기',
+                      suffixIcon: Icon(Icons.search),
+                    ),
                   ),
-                ),
-                SizedBox(height: 5),
-                _tip(),
-              SizedBox(height: 5), ...
-              _classList(),
-                CustomButton(
-                  label: "수강 신청하기",
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(builder: (context) => registation()),
-                    );
-                  },
-                ),
-              ],
+                  SizedBox(height: 5),
+                  
+                  Tip(),
+
+                SizedBox(height: 5), ...
+                _classList(),
+                  CustomButton(
+                    label: "수강 신청하기",
+                    onPressed: () {
+                      Navigator.push(
+                        context,
+                        MaterialPageRoute(builder: (context) => Registation()),
+                      );
+                    },
+                  ),
+                ],
+              ),
             ),
           ),
         ),
-      ),
-    );
+      );
   }
 
   /// 수강하는 과목들이 list로 보이고,
@@ -75,15 +76,7 @@ class classList extends StatelessWidget {
   }
 
   /// 검색하기에서 초성 검색도 가능하다는 tip 문구
-  Widget _tip() {
-    return const Align(
-              alignment: Alignment.centerLeft,
-              child: Text(
-                'tip. 초성 검색도 가능해요 ex) ㅁㄹㅎ > 물리학',
-                style: TextStyle(color: Colors.white, fontSize: 14),
-                textAlign: TextAlign.left,
-              ),);
-  }
+
 
   AppBar _appbar(BuildContext context) {
     return AppBar(
